Readme.sh network availability test, and bug fix in the sce-update routine.

+++ usr/bin/readme.sh
@@ -6,6 +6,9 @@ export PATH
 BUILD=`getBuild`
 MIRROR=`cat /opt/tcemirror`
 README="$1"
+checknotroot
+[ -f /tmp/select.ans ] && sudo rm /tmp/select.ans
+
 if [ "$1" == "-h" ] || [ "$1" == "--help" ]; then
 	echo " "
 	echo "${YELLOW}readme.sh - View README files for dCore packages.${NORMAL}"
@@ -17,10 +20,26 @@ if [ "$1" == "-h" ] || [ "$1" == "--help" ]; then
 	echo " "
 fi
 
+exit_tcnet() {
+	echo "There is an issue connecting to "$MIRROR", exiting.."
+	exit 1
+}
+
+if /bb/wget -s "$MIRROR" > /dev/null 2>&1; then
+	:
+else
+	exit_tcnet
+fi
+
 if [ -z "$1" ]; then
 	wget -q "$MIRROR"/dCore/"$BUILD"/README/READMELIST -O /tmp/READMELIST
 	select2 "Select README file." /tmp/READMELIST	
 	README="$(cat  /tmp/select.ans)"
+fi
+
+if [ -z "$README" ]; then
+	echo "No selection made, exiting.."
+	exit 0
 fi
 
 [ -f /tmp/README-"$README".txt ] && rm /tmp/README-"$README".txt
